+++ .gitignore
@@ -1,6 +1,7 @@
 .classpath
 .project
 .settings/
+.springBeans
 target
 
 .idea
