+++ function-controller/src/main/java/io/projectriff/controller/function/SidecarProperties.java
@@ -1,0 +1,36 @@
+/*
+ * Copyright 2017 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.projectriff.controller.function;
+
+import org.springframework.boot.context.properties.ConfigurationProperties;
+
+/**
+ * @author Mark Fisher
+ */
+@ConfigurationProperties(prefix = "riff.function.sidecar")
+public class SidecarProperties {
+
+	private String tag = "0.0.1-SNAPSHOT";
+
+	public String getTag() {
+		return tag;
+	}
+
+	public void setTag(String tag) {
+		this.tag = tag;
+	}
+}
