+++ function-controller/src/main/resources/application.properties
@@ -1,1 +1,1 @@
-logging.level.io.sk8s.function.controller = trace
+logging.level.io.projectriff.controller.function = trace
