+++ riff-core/src/main/java/io/projectriff/core/resource/ResourceAddedEvent.java
@@ -1,0 +1,31 @@
+/*
+ * Copyright 2017 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.projectriff.core.resource;
+
+import io.fabric8.kubernetes.client.Watcher;
+
+/**
+ * Specialization of {@link ResourceEvent} for when a resource is added.
+ * @author Eric Bottard
+ * @param <T> the type of resource that is being added
+ */
+public class ResourceAddedEvent<T> extends ResourceAddedOrModifiedEvent<T> {
+
+	ResourceAddedEvent(T resource) {
+		super(resource, Watcher.Action.ADDED);
+	}
+}
