--- riff-test-support/src/main/java/io/sk8s/test/KafkaAvailableRule.java
@@ -1,44 +1,0 @@
-/*
- * Copyright 2017 the original author or authors.
- *
- *   Licensed under the Apache License, Version 2.0 (the "License");
- *   you may not use this file except in compliance with the License.
- *   You may obtain a copy of the License at
- *
- *        http://www.apache.org/licenses/LICENSE-2.0
- *
- *   Unless required by applicable law or agreed to in writing, software
- *   distributed under the License is distributed on an "AS IS" BASIS,
- *   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *   See the License for the specific language governing permissions and
- *   limitations under the License.
- */
-
-package io.sk8s.test;
-
-import static org.assertj.core.api.Assertions.assertThat;
-
-/**
- * @author David Turanski
- **/
-public class KafkaAvailableRule extends KubernetesAvailableRule {
-	private final static String KAFKA_LABEL = "app=kafka";
-
-	public KafkaAvailableRule() {
-		super("K8S KAFKA");
-	}
-
-	@Override
-	protected void obtainResource() throws Exception {
-		try {
-			boolean serviceAvailable = !this.client.services().withLabel(KAFKA_LABEL).list().getItems().isEmpty();
-			boolean podsAvailable = !this.client.pods().withLabel(KAFKA_LABEL).list().getItems().isEmpty();
-
-			assertThat(serviceAvailable && podsAvailable).isTrue();
-		}
-		catch (Error e) {
-			throw new RuntimeException("Kafka on k8s is not available.");
-		}
-	}
-
-}
