--- riff-test-support/src/main/java/io/sk8s/test/Sk8sTypesAvailableRule.java
@@ -1,40 +1,0 @@
-/*
- * Copyright 2017 the original author or authors.
- *
- *   Licensed under the Apache License, Version 2.0 (the "License");
- *   you may not use this file except in compliance with the License.
- *   You may obtain a copy of the License at
- *
- *        http://www.apache.org/licenses/LICENSE-2.0
- *
- *   Unless required by applicable law or agreed to in writing, software
- *   distributed under the License is distributed on an "AS IS" BASIS,
- *   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *   See the License for the specific language governing permissions and
- *   limitations under the License.
- */
-
-package io.sk8s.test;
-
-import io.projectriff.kubernetes.client.RiffClient;
-
-/**
- * @author David Turanski
- **/
-public class Sk8sTypesAvailableRule extends KubernetesAvailableRule {
-
-	public Sk8sTypesAvailableRule() {
-		super("Sk8s TYPES");
-	}
-
-	@Override
-	protected void obtainResource() throws Exception {
-		try {
-			this.client.adapt(RiffClient.class).functions().list();
-			this.client.adapt(RiffClient.class).topics().list();
-		}
-		catch (Exception e) {
-			throw new RuntimeException("sk8s types are not available.");
-		}
-	}
-}
